feat: manifest

+++ lsm/table.go
@@ -13,7 +13,7 @@ import (
 type table struct {
 	ss   *file.SSTable
 	lm   *levelManager
-	fid  uint32
+	fid  uint64
 	idxs []byte
 }
